refactor: InfoInputSection 및 InfoInputContainer 컴포넌트명 변경
