+++ src/feature/CardRegister/components/CardRegisterForm/CardRegisterForm.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import { fn } from "storybook/test";
 import { MemoryRouter } from "react-router-dom";
 
-import InfoInputSection from "./InfoInputSection";
+import InfoInputSection from "./CardRegisterForm";
 import type {
   CardInfoType,
   CardNumberChunkType,
