+++ src/feature/CardRegister/components/CardRegisterForm/CardRegisterStep/CardRegisterStep.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
 import Input from "../../../../../common/components/Input/Input";
-import InputContainer from "./InputContainer";
+import InputContainer from "./CardRegisterStep";
 
 const meta = {
   title: "feature/CardRegister/components/InputContainer",
